+++ packages/component-events/src/components/CardsGridEvents/index.jsx
@@ -15,7 +15,13 @@ import { GridView } from "./GridView";
  * @param {FeedType} props
  * @returns {JSX.Element}
  */
-const CardsGridEvents = ({ header, ctaButton, dataSource, maxItems }) => {
+const CardsGridEvents = ({
+  header,
+  ctaButton,
+  dataSource,
+  maxItems,
+  noFeedText = "No events to show.",
+}) => {
   useEffect(() => {
     if (typeof window !== "undefined") {
       trackReactComponent({
@@ -33,7 +39,7 @@ const CardsGridEvents = ({ header, ctaButton, dataSource, maxItems }) => {
   }, []);
   return (
     // Calling the high order component that fetch the data
-    <BaseFeed {...{ header, ctaButton, dataSource, maxItems }}>
+    <BaseFeed {...{ header, ctaButton, dataSource, maxItems, noFeedText }}>
       <GridView />
     </BaseFeed>
   );